+++ tests/testthat/test-precipitation_lower.R
@@ -8,7 +8,7 @@ library(data.table)
 context("rr transformation (lower) - Entire station set")
 
 subscenario    <- "lower"
-ifile      <- "regressionInput/precipitation/KNMI14____ref_rrlower___19810101-20101231_v3.2.txt"
+ifile      <- "regressionInput/precipitation/KNMI14____ref_rrcentr___19810101-20101231_v3.2.txt"
 ofile      <- NA
 
 test_that("2030 decadal prediction", {